+++ flexmeasures/api/v3_0/accounts.py
@@ -36,8 +36,7 @@ class AccountAPI(FlaskView):
         .. :quickref: Account; Download account list
 
         This endpoint returns all accessible accounts.
-        Accessible accounts are your own account, the accounts you are a consultant for, or all accounts for admins.
-        When the super-account concept (GH#203) lands, then users in such accounts see all managed accounts.
+        Accessible accounts are your own account and accounts you are a consultant for, or all accounts for admins.
 
         **Example response**
 
@@ -67,8 +66,8 @@ class AccountAPI(FlaskView):
             accounts = get_accounts()
         else:
             accounts = [current_user.account] + (
-                current_user.account.consultant_client_accounts
-                if "customer-manager" in current_user.roles
+                current_user.account.consultancy_client_accounts
+                if "consultant" in current_user.roles
                 else []
             )
 
@@ -84,7 +83,7 @@ class AccountAPI(FlaskView):
         .. :quickref: Account; Get an account
 
         This endpoint retrieves an account, given its id.
-        Only admins, consultant users or the user themselves can use this endpoint.
+        Only admins, consultants and users belonging to the account itself can use this endpoint.
 
         **Example response**
 
@@ -94,8 +93,7 @@ class AccountAPI(FlaskView):
                 'id': 1,
                 'name': 'Test Account'
                 'account_roles': [1, 3],
-                'consultancy_account_id':2,
-                'consultant_name':'Consultant',
+                'consultancy_account_id': 2,
             }
 
         :reqheader Authorization: The authentication token